refactor: extract changelog trim logic into independent script
- Move trimChangelogs() from afterPack.js to new scripts/trimChangelogs.js
- afterPack.js now imports and delegates to the independent module
- Dockerfile calls the independent script instead of afterPack.js
